+++ CHANGELOG.md
@@ -12,6 +12,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - ![API] Add warming of network resources [#919](https://github.com/nbbrd/sdmx-dl/issues/919)
 - ![API] Add dimension code availability API [#1016](https://github.com/nbbrd/sdmx-dl/issues/1016)
 - ![API] Add convenient request API [#1027](https://github.com/nbbrd/sdmx-dl/issues/1027)
+- ![API] Add short representation of ResourceRef
 - ![SOURCE] Add sources from Canadian Center of Energy Information [#1055](https://github.com/nbbrd/sdmx-dl/issues/1055)
 - ![SOURCE] Add source from Canadian Census Profile Data [#1056](https://github.com/nbbrd/sdmx-dl/issues/1056)
 - ![GRPC] Add dimension code availability in protobuf service [#912](https://github.com/nbbrd/sdmx-dl/issues/912)